Add files via upload
This notebook will pull in a starting airfoil geometry "naca0012" and open the xfoil.exe program input a range of angle of attack and Reynolds number. If the program times out there is a negative reward and move to the next simulation. Reward is based off the average cl/cd over angles of attack 15-25. Reward is also calculated as that value-previous cl/cd in order to reward improvements. This environment successfully runs xfoil as desired, however there is no machine learning or application of reward. Currently working to implement this feature.